refactor(core): move SubRequest and SubResponse to core crate
SubRequest and SubResponse are pure data types with no
filter-specific dependencies. Move them from filter/src/pipeline/subrequest.rs
to core/src/subrequest.rs alongside SubRequestConnector, preserving the
filter crate's public API via re-exports.
