+++ core/Cargo.toml
@@ -17,7 +17,9 @@ name = "praxis_core"
 workspace = true
 
 [dependencies]
+bytes = { workspace = true }
 dashmap = { workspace = true }
+http = { workspace = true }
 pingora-core = { workspace = true }
 praxis-tls = { workspace = true }
 rand = { workspace = true }
